+++ docs/src/SubNav.js
@@ -53,10 +53,10 @@ class SubNav extends React.Component {
       return true;
     }
 
-    if (props.children) {
-      return ValidComponentChildren.some(props.children, child => (
-        this.isActive(child, activeKey, activeHref)
-      ));
+    if (ValidComponentChildren.some(props.children, (child) => (
+      this.isActive(child, activeKey, activeHref)
+    ))) {
+      return true;
     }
 
     return props.active;
@@ -72,7 +72,7 @@ class SubNav extends React.Component {
       className,
       style,
       children,
-      ...props,
+      ...props
     } = this.props;
     const [, elementProps] = splitBsProps(props);