+++ src/FormControl.js
@@ -5,7 +5,9 @@ import warning from 'warning';
 
 import FormControlFeedback from './FormControlFeedback';
 import FormControlStatic from './FormControlStatic';
-import { bsClass, getClassSet, splitBsProps } from './utils/bootstrapUtils';
+import { bsClass, getClassSet, splitBsProps, bsSizes } from './utils/bootstrapUtils';
+import { SIZE_MAP, Size } from './utils/StyleConfig';
+import { prefix } from './utils/bootstrapUtils';
 
 const propTypes = {
   componentClass: elementType,
@@ -17,6 +19,14 @@ const propTypes = {
    * Uses `controlId` from `<FormGroup>` if not explicitly specified.
    */
   id: React.PropTypes.string,
+  /**
+   * Attaches a ref to the `<input>` element. Only functions can be used here.
+   *
+   * ```js
+   * <FormControl inputRef={ref => { this.input = ref; }} />
+   * ```
+   */
+  inputRef: React.PropTypes.func,
 };
 
 const defaultProps = {
@@ -36,8 +46,10 @@ class FormControl extends React.Component {
       componentClass: Component,
       type,
       id = controlId,
+      inputRef,
       className,
-      ...props,
+      bsSize,
+      ...props
     } = this.props;
 
     const [bsProps, elementProps] = splitBsProps(props);
@@ -53,11 +65,19 @@ class FormControl extends React.Component {
       classes = getClassSet(bsProps);
     }
 
+    // If user provides a size, make sure to append it to classes as input-
+    // e.g. if bsSize is small, it will append input-sm
+    if (bsSize) {
+      const size = SIZE_MAP[bsSize] || bsSize;
+      classes[prefix({ bsClass: 'input' }, size)] = true;
+    }
+
     return (
       <Component
         {...elementProps}
         type={type}
         id={id}
+        ref={inputRef}
         className={classNames(className, classes)}
       />
     );
@@ -71,4 +91,6 @@ FormControl.contextTypes = contextTypes;
 FormControl.Feedback = FormControlFeedback;
 FormControl.Static = FormControlStatic;
 
-export default bsClass('form-control', FormControl);
+export default bsClass('form-control',
+  bsSizes([Size.SMALL, Size.LARGE], FormControl)
+);