+++ src/Nav.js
@@ -93,6 +93,7 @@ const defaultProps = {
 const contextTypes = {
   $bs_navbar: React.PropTypes.shape({
     bsClass: React.PropTypes.string,
+    onSelect: React.PropTypes.func,
   }),
 
   $bs_tabContainer: React.PropTypes.shape({
@@ -271,7 +272,7 @@ class Nav extends React.Component {
       pullLeft,
       className,
       children,
-      ...props,
+      ...props
     } = this.props;
 
     const tabContainer = this.context.$bs_tabContainer;
@@ -319,6 +320,7 @@ class Nav extends React.Component {
           const childOnSelect = createChainedFunction(
             child.props.onSelect,
             onSelect,
+            navbar && navbar.onSelect,
             tabContainer && tabContainer.onSelect
           );