+++ src/NavDropdown.js
@@ -31,10 +31,10 @@ class NavDropdown extends React.Component {
       return true;
     }
 
-    if (props.children) {
-      return ValidComponentChildren.some(props.children, child => (
-        this.isActive(child, activeKey, activeHref)
-      ));
+    if (ValidComponentChildren.some(props.children, (child) => (
+      this.isActive(child, activeKey, activeHref)
+    ))) {
+      return true;
     }
 
     return props.active;
@@ -48,7 +48,7 @@ class NavDropdown extends React.Component {
       className,
       style,
       children,
-      ...props,
+      ...props
     } = this.props;
     const [bsProps, elementProps] = splitBsProps(props);