+++ src/Tabs.js
@@ -45,6 +45,11 @@ const propTypes = {
   onSelect: React.PropTypes.func,
 
   /**
+   * Wait until the first "enter" transition to mount tabs (add them to the DOM)
+   */
+  mountOnEnter: React.PropTypes.bool,
+
+  /**
    * Unmount tabs (remove it from the DOM) when it is no longer visible
    */
   unmountOnExit: React.PropTypes.bool,
@@ -53,7 +58,8 @@ const propTypes = {
 const defaultProps = {
   bsStyle: 'tabs',
   animation: true,
-  unmountOnExit: false
+  mountOnEnter: false,
+  unmountOnExit: false,
 };
 
 function getDefaultActiveKey(children) {
@@ -90,13 +96,14 @@ class Tabs extends React.Component {
       id,
       onSelect,
       animation,
+      mountOnEnter,
       unmountOnExit,
       bsClass,
       className,
       style,
       children,
       activeKey = getDefaultActiveKey(children),
-      ...props,
+      ...props
     } = this.props;
 
     return (
@@ -119,6 +126,7 @@ class Tabs extends React.Component {
           <TabContent
             bsClass={bsClass}
             animation={animation}
+            mountOnEnter={mountOnEnter}
             unmountOnExit={unmountOnExit}
           >
             {children}