+++ test/ModalTitleSpec.js
@@ -2,47 +2,48 @@ import React from 'react';
 import ReactTestUtils from 'react-addons-test-utils';
 import ReactDOM from 'react-dom';
 
-import Media from '../src/Media';
+import Modal from '../src/Modal';
 
-describe('Media.Heading', () => {
+describe('Modal.Title', () => {
   it('uses "h4" by default', () => {
     const instance = ReactTestUtils.renderIntoDocument(
-      <Media.Heading />
+      <Modal.Title />
     );
 
     assert.equal(ReactDOM.findDOMNode(instance).nodeName, 'H4');
   });
 
-  it('has "media-heading" class', () => {
+  it('has "modal-title" class', () => {
     const instance = ReactTestUtils.renderIntoDocument(
-      <Media.Heading />
+      <Modal.Title />
     );
 
-    assert.include(ReactDOM.findDOMNode(instance).className, 'media-heading');
+    assert.include(ReactDOM.findDOMNode(instance).className, 'modal-title');
   });
 
   it('should merge additional classes passed in', () => {
     const instance = ReactTestUtils.renderIntoDocument(
-      <Media.Heading className="custom-class" />
+      <Modal.Title className="custom-class" />
     );
+    const classes = ReactDOM.findDOMNode(instance).className;
 
-    assert.include(ReactDOM.findDOMNode(instance).className, 'media-heading');
-    assert.include(ReactDOM.findDOMNode(instance).className, 'custom-class');
+    assert.include(classes, 'modal-title');
+    assert.include(classes, 'custom-class');
   });
 
   it('should allow custom elements instead of "h4"', () => {
     const instance = ReactTestUtils.renderIntoDocument(
-      <Media.Heading componentClass="h2" />
+      <Modal.Title componentClass="h3" />
     );
 
-    assert.equal(ReactDOM.findDOMNode(instance).nodeName, 'H2');
+    assert.equal(ReactDOM.findDOMNode(instance).nodeName, 'H3');
   });
 
   it('should render children', () => {
     const instance = ReactTestUtils.renderIntoDocument(
-      <Media.Heading>
+      <Modal.Title>
         <strong>Children</strong>
-      </Media.Heading>
+      </Modal.Title>
     );
     assert.ok(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'strong'));
   });