+++ test/NavDropdownSpec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactTestUtils from 'react/lib/ReactTestUtils';
+import ReactTestUtils from 'react-addons-test-utils';
 import ReactDOM from 'react-dom';
 
 import MenuItem from '../src/MenuItem';
@@ -103,6 +103,27 @@ describe('<NavDropdown>', () => {
     expect(ReactDOM.findDOMNode(items[0]).className).to.not.match(/active/);
     expect(ReactDOM.findDOMNode(items[1]).className).to.match(/active/);
     expect(ReactDOM.findDOMNode(items[2]).className).to.not.match(/active/);
+  });
+
+  it('should handle nested child null active state', () => {
+    class Container extends React.Component {
+      render() {
+        return null;
+      }
+    }
+
+    const instance = ReactTestUtils.renderIntoDocument(
+      <NavDropdown id="test-id" title="title">
+        <Container>
+          <MenuItem>MenuItem 1 content</MenuItem>
+        </Container>
+      </NavDropdown>
+    );
+
+    const container = ReactTestUtils.findRenderedComponentWithType(
+      instance, Container,
+    );
+    expect(container.props.active).to.not.be.false;
   });
 
   it('should derive bsClass from parent', () => {